A temporary fix for the Case of Runaway Jade.
DarkGod, space kerns with some letters. Kerns heavily. SDL_TTF also
sucks heavily enough to not tell us that. If you could change font:draw
so that it collected all the text in the line (of one style at least)
and rendered it at once, thus preserving kerning, that'd be wonderful.
Again, when we move to atlasi, this is going to become a pain again.
Probably will need something which can return full character metrics.