Fix Rune: Biting Gale not being used by tactical AI
This was caused by the lack of a "radius" attribute.

Which tactical.lua makes good use of the talent's "target" value, for
the initial range check it ignores this and uses the sum of
getTalentRadius and getTalentRange. So, while Biting Gale had a
perfectly good target table, because the radius was not stored in a
separate radius value the AI was assuming it to have a range of 0 and
thus never using it.

I've searched through the entire code base, and not found any other
cases exactly like this. I have, however, found a few which could be
mildly improved. I'll consider uploading a bundle of such fixes later.