Character sheet: The default frame size is based on the size of the font (200*font_w, 36*font_h) to maintain readability at different screen resolutions (tested from 800x600 to 1920x1200). Added an Equipment tab (with character doll) that allows you to inspect equipment (including offsets) of actors while inspecting them. Added a button to open the equipinven dialog (for the player only) which triggers the display updates, making equipment/stat management much easier and faster. Added hotkeys to allow direct select of each tab, and the initially displayed tab can be specified when the dialog is called. Redefined the default key binding for shift-c to open the character sheet for the actor under the mouse or within the targetting reticle if possible.  This allows targeted enemies to be inspected during the targeting process. Added/enhanced tooltips for several items to provide more explanation of game mechanics. Fixed the display to always show comparison changes to resists/cap (including actor-specific), damage affinities, status immunities, damage/damage penetration bonuses (including actor-specific). Fixed a formatting issue that was causing some talents to be cut off on the talent tab.
Fixed a bug that can cause ShowEquipInven to crash for actors with no defined main inventory (which is defined as required with default parameters).

EquipDoll:
Added support for a new scaling option that will resize the doll frame, graphics, and fonts (fonts are assumed to all be truetype).
A specific/custom frame definition can be specified during instanciation for on-the-fly EquipDoll generation.
The exact fonts used can be specified (which allows/requires manual scaling).
The title can be specified or disabled during instaciation.
The size of the doll graphic can be specified and the frame will be correctly sized for it if it is enabled.
Added a new "abovetop" text position option to EquipDollFrames to allow for staggering text on a line.

Objects:
Updated getTextualDesc to provide more information for unidentified objects including the presence of  special damage types inflicted and special effects, and the presence of activated abilities.
Revised getShortName to note ego level instead of the list of keywords for unidentified objects.