Support multiple resources for Arcane Combat
-The only visible change in the core game is that Arcane Combat
now interacts correctly with the Hidden Resources prodigy. The rest
of the changes are for addon support.

-Previously, Arcane Combat on_trigger function assumed that allowed
talents cost mana and not other resources. This is fine in the core
game, but if an addon wants to allow triggering a talent that costs
Vim or Paradox, for example, the logic was not appropriate.
Now, all resource types should be handled correctly.

-Added some sanity checks to the allowed talent filtering in case
addon developers set "allow_for_arcane_combat" on inappropriate
talents.