Restructure cloning code and more
1. Since the customizable table cloning function may be useful for
other purposes than just making Chronomancy clones, I moved it to
engine/class.lua and renamed it cloneCustom(). It works like
cloneFull(), but allows specifying a table of keys/values to skip or
replace in the clone table.

If desired, I could instead just modify cloneFull() to have an optional
parameter for the "alt_nodes" table, but I didn't want to mess with it
without permission. There could be a conditional statement to pick
between two different recursive functions so the optional param doesn't
impact performance, if that's a concern.

2. Updated comments for cloneFull() and related functions for clarity.
The comment originally said it returns the clone and the number of
objects cloned, but it actually only returns the clone. I think the
comment may have been copied from somewhere else and not updated.

3. Renamed the "t" param from cloneFull() to "post_copy" to be more
descriptive and to be consistent with cloneCustom(). This has no
functional impact.

4. Significantly expanded the number of properties which are skipped
when making Chronomancy clones of an actor. Hopefully this will improve
performance and maybe even help with the memory leak. I didn't see any
errors during brief testing, but it probably merits more testing.