Add Arrow Stitching cleanup function and more
Added a cleanup function for Arrow Stitching, since it spawns clones
the same way as Warden's Call. The cleanup function is the same.

Changed makeParadoxClone() to not copy fov and distance_map tables from
the source actor. Instead, we call ActorFOV.init() on the clone.

Updated a few comments for clarity.