Fixes issues with major events (that create an additional level): sub-vault, rat-lich, damp-cave, old-battle-field, drake-cave, naga-portal, fearscape-portal
.no_worldport is set to that of the parent zone.
Actor levels are 5% to 15% higher than the the parent zone and game difficulty adjustments are not compounded.
Object material levels match the parent zone.
Auto explore and minimap settings updated.

Fixed a bug in sub-vault that was preventing the stairs from crumbling as intended.

The player can choose to ignore the naga-portal and fearscape-portal by hitting escape at the pop-up.

Fixed a bug in game:changeLevelReal that was causing the player to incorrectly be placed at level start when params.auto_level_stair was set.

The player is correctly placed on one of the graves when returning from the old-battle-field event.