GameState:doneEvent updated to allow incrementing the used count (which is returned)
GameState:startEvents() updated:
The zone events file supports some new fields:
unique = true prevents triggering of events that have previously been triggered (anywhere in the game)
level_range = a range of allowed levels the event is allowed to spawn on (complements the forbid field)
special = an optional function(lev) that can determine if the event should spawn or not
(Logic reversed: must now return true to allow the event to trigger, consistent with other "special" functions)
doneEvent is always triggered for events that are placed or assigned to a level

Updated the sludgenest and noxious-caldera events for the wilderness and the conclave-vault event for the halfling ruins.  (They special field

is removed as it's not needed.)

Added documentation and output to the relevant GameState functions.

Added (with appropriate level restrictions) the rat-lich, old-battle-field, damp-cave, drake-cave, fearscape-portal, and naga-portal events to

the I.D. The cultists event in the I.D. is now unique and can only spawn from level 5+.