Updated Actor.on_set_temporary_effect to use Combat:checkHitOld instead of Combat:checkHit, which uses a broader curve when making saves for
temporary effects.  This results in saves for temporary effects (when .apply_power is set) always having a chance both to succeed or to fail,

and bonuses to save values are always useful (more is always better) regardless of the power of the attack saved against.

In addition, if a save is failed, there is always a chance to reduce the duration of the effect, even if the save is less than the applied

power.  The expected duration on a failed save steadily decreases as a save increases versus the power applied.  (This is an evolution of the

changess discussed in http://forums.te4.org/viewtopic.php?f=36&t=36715.)  The effect of saves on durations can be easily tuned by modifying the

variables mean_fact and std_dev as documented in the code.

Overall, these changes are a modest buff to saves against temporary effects, while making them less reliable at high values.

This should help curb chain CC issues that can happen against higher level NPCs, since not all effects will land with full duration, while

helping prevent effective immunity for the same NPCs.

Additional Fixes:
The duration of an existing effect will not be reduced by setting a new effect with a shorter duration (controlled by on_merge for some

effects).  This means that an effect will not be removed if you save against another application of the same effect.