Updated Combat:combatDamage to add total weapon stat bonuses to the combatDamagePower (affecting the multiplier).
The effect is to curb runaway scaling of very high stat bonuses (e.g. Temporal Hounds, etc.) while restoring most combat damage to previous levels.

Added an extra line in the Character sheet (active when debug mode is enabled) showing the original combat damage for comparison.  (This is intended to be temporary -- remove the marked code after rescaling of melee damage is set.)