Disable Dreamscape on Dreaming Horrors
Something is very bugged with this talent, nobody knows what exactly it is, but it can effect future characters as well as current so it needs to be disabled for now.

Its possible hachem_muche's changes to cloning have fully fixed this issue, unfortunately I don't know how to test that.